INSTALLATION.txt - Updated the Trilinos installation instructions.
fipy/terms/convectionTerm.py - changed the logic of how to calculate the diffusion coefficient if it is zero.

examples/convection/source.py - added an example to demonstrate a ConvectionTerm with a source term. This example tests a previously checked in change that fixed a bug in the 
default sign of the diffusion coefficient when there is no diffusion coefficient in the equation. 



git-svn-id: svn+ssh://code.matforge.org/fipy/trunk@2927 d80e17d7-ff13-0410-a124-85740d801063