ticket:462 -- fix for broken tests on loki -- Epetra SumAll broken
Tests in `fipy.meshes.gmshMesh.Gmsh2D` were failing because for a
serial job `self.epetra_comm.SumAll(numerix.array(a).sum(axis=axis))`
only returns an array of shape (), which fails with `TypeError:
Require int, long or double array, got notype array` for some versions
of Trilinos. Now passes an array of shape (1,) if shape is () and
reshapes back to () on exit.

This issue occurs with:
{{{
    mpi4py version 1.2.2
    Trilinos version: 10.4.0
    PyTrilinos version: 4.4
    numpy version 1.6.1
}}}
on `x86_64`.