Fix for ticket:465, using diagonalSign with convection terms.
A `_getDiagonalSign` method is now in `nonDiffusionTerm.py` and is
used by both `ImplicitSouceTerm` and `ConvectionTerm`'s to inform them
of the sign of the equation. `ConvectionTerm`'s need to know the sign
of the equation so that the upwind direction can be evaluated
correctly when the diffusion term is absent.

 * Removing stray import in `fipy/terms/cellTerm.py`.

 * Adding test case that tests the solution to an equation is
   independent of sign.

 * Setting the default diffusion coefficient to be small in comparison
   to the convection coefficient rather than using an absolute small
   value.