Created unifying interface for LSMLIB and Scikit-fmm.
Both LSMLIB and Scikit-fmm have been modified. Two tests fail for
Scikit-fmm. This is probably due to subtle differences in the
calculation of the distance function for second order. It is related to
differences caused in the order that the updates take place for trial
values which have the same value. The level set algorithm is not
entirely deterministic.

 * Recreated LSMLIB doctest directive for tests that only pass with
   LSMLIB.