Adding a test for a bug in both LSMLIB and Scikit-fmm.
The test gives different results depending on whether LSMLIB or
Scikit-fmm_ is used. There is a deeper problem that is related to this
issue. When a value becomes "known" after previously being a "trial"
value it updates its neighbors' down winding values, but it fails to
update other downwind values when using second order differencing.