Using `Popen(shell=True)` rather than `shell=False` for launching Gmsh
This is needed as part of a test to get a minimal Buildbot environment
configured. For some reason when `shell=False`, the local `gmsh`
executable in the virtualenv from which the script is launched is not
used, but the version in `/usr/bin` is used instead. There are some
risks with using `shell=True`, but I want to see what Buildbot does.