Fixing ticket:490, renaming GridXD to be NonUniformGridXD.
This is part of the resolution for ticket:490. Avoids name clashes
between the grid factory functions and identically named classes. The
GridXD and CylindricalGridXD classes are now deprecated. Obviously,
the factory functions by the same name have not been changed. This
fixes the issue of importing the wrong GridXD accidentally as happened
in the email that prompted ticket:490.

 * Deprecation warnings have been added in the gridXD.py files.