Dissentagle comms from solvers
Old `parallelComm` made heavy assumptions about availability of Trilinos.
Comms (particularly the parallel ones) are intimately tied to the solver
in use, so they should be instantiated with the solver.

No matter what, we need mpi4py, so use it (and only it) to determine number
of processes in order to pick a solver.

Addresses #644