Bug with Matplotliv 1.4.0 is fixed
Fix ticket:672

Bug with Matplotlib version 1.4.0 is fixed. the bug was caused because
the default value of vmin and vmax is None internally to
Matplotlib. Some code in matplotlib tries a numpy operation on these
None values. A simple fix is to pass a norm instance to the
ColorbarBase class when it is called in FiPy. This may be an internal
bug to Matplotlib.

The fix works was tried with both Matplotlib version 1.4.0 and 1.3.1.