Clear pdflatex build errors by removing :term:`Python` from heading
Using :term:`Python` Programming Language as a header in design.rst
produces incomplete/erroneous \def's in fipy.toc. pdflatex then
requests user intervention to clear the error. The crude fix is to
remove the :term: tag from the heading. The root cause may be in
Sphinx, or there may be an elegant way to hide the restructured
text markup from makeindex.