Memorize the C string representation.
For some cases, the computation of the C
string for inline using _getRepresentation
takes a substantial amount of time.

Here, we memorize the computed C code and argument dictionary.
The argDict used to store the values of the required variables.
Now it stores the variables themselves,
and the values are retrieved by _execInline.