Remove numerix.indices
This routine is redundant to `numpy.indices` and raises an `Exception`
on processes with no cells, which deadlocks. `numpy.indices` doesn't
suffer the same fate.

This was introduced in 2006 (commit b0758f55) as part of
"Improving UniforGrid2D efficiency with the use of inlining".
Removing it seems to do no harm, with or without inlining.