Always run the cron job
The midnight build is scheduled precisely to catch any changes in the test environment. It hasn’t run since November 9th. "By
default, your pipeline doesn't run as scheduled if there have been no code changes since the last successful scheduled run.” [1]

Grrr!!! We have code change triggers to build when the code changes. The whole point of a scheduled build is that it runs on a flipping
schedule!

[1]: https://learn.microsoft.com/en-us/azure/devops/pipelines/process/scheduled-triggers?view=azure-devops&tabs=yaml#running-even-when-there-are-no-code-changes