Consolidate Solver default parameters
Eliminate cumbersome telephone game for a small number of
changed defaults in subclasses. Previous was error-prone
and tedious to change global defaults.

Contrary to
[official guidance](https://docs.python-guide.org/writing/gotchas/#what-you-should-do-instead),
use `"default"` instead of `None` to denote default argument.
Semantics of using None are confusing when there are times we want
no preconditioner or no tolerance limit and other times we want
a default preconditioner or default tolerance limit.